requirements.txt
ipython==9.1.0
black==25.1.0
pylint==3.3.6
pytest==8.3.5
huggingface-hub==0.30.2
python-dotenv==1.1.0
markdownify==1.1.0
smolagents==1.13.0
requests==2.32.3
duckduckgo_search==8.0.1
pandas==2.2.3
gradio==5.25.2
langchain==0.3.24
langchain-community==0.3.22
rank-bm25==0.2.2
helium==5.1.1
langgraph==0.3.34
langchain-openai==0.3.14
langchain-huggingface==0.1.2
langchain-google-genai==2.1.3
langchain-core==0.3.55
langfuse==2.60.3
llama-index==0.12.33
llama-index-llms-huggingface-api==0.4.1
llama-index-retrievers-bm25==0.5.2
llama-index-tools-duckduckgo==0.3.0
bitsandbytes
peft
trl
tensorboardX
wandb
smolagents[telemetry]
opentelemetry-sdk
opentelemetry-exporter-otlp
openinference-instrumentation-smolagents
langfuse
datasets
smolagents[gradio]
